requirements.txt
fastapi
uvicorn[standard]
peewee
pymysql
cryptography
python-dotenv
requests
python-multipart
pyjwt


